+++ src/main/java/br/com/jorgemelo/nimbusfilemanager/duplicate/application/VideoSimilarityAsyncRunner.java
@@ -42,7 +42,7 @@ public class VideoSimilarityAsyncRunner {
 		return runner.total();
 	}
 
-	public int percent() {
+	public double percent() {
 		return runner.percent();
 	}
 }
