+++ src/main/java/br/com/jorgemelo/nimbusfilemanager/geolocation/application/dto/Snapshot.java
@@ -2,7 +2,7 @@ package br.com.jorgemelo.nimbusfilemanager.geolocation.application.dto;
 
 import br.com.jorgemelo.nimbusfilemanager.geolocation.domain.enums.Phase;
 
-public record Snapshot(Phase phase, String stepLabel, long bytesDone, long bytesTotal, int percent, long etaSeconds,
+public record Snapshot(Phase phase, String stepLabel, long bytesDone, long bytesTotal, double percent, long etaSeconds,
 		long recordsImported) {
 
 	public boolean downloading() {
